// packages/isaac-typescript-definitions/src/enums/DoorSlot.d.ts
export declare enum DoorSlot {
    NO_DOOR_SLOT = -1,
    /** In a normal room, the left door. In a 2x2 room, the uppermost door on the left side. */
    LEFT_0 = 0,
    /** In a normal room, the top door. In a 2x2 room, the leftmost door on the top side. */
    UP_0 = 1,
    /** In a normal room, the right door. In a 2x2 room, the uppermost door on the right side. */
    RIGHT_0 = 2,
    /** In a normal room, the bottom door. In a 2x2 room, the leftmost door on the bottom side. */
    DOWN_0 = 3,
    /** In a 2x2 room, the bottommost door on the left side. */
    LEFT_1 = 4,
    /** In a 2x2 room, the rightmost door on the top side. */
    UP_1 = 5,
    /** In a 2x2 room, the bottommost door on the right side. */
    RIGHT_1 = 6,
    /** In a 2x2 room, the rightmost door on the bottom side. */
    DOWN_1 = 7
}